test: http2 compat response.write() error checks
PR-URL: https://github.com/nodejs/node/pull/18859
